backend/requirements.txt
# Core Framework
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6

# Document Processing
pymupdf==1.23.0
pytesseract==0.3.10
pillow==10.1.0
pdf2image==1.16.3

# Smart Text Processing
spacy==3.7.0
sentence-transformers==2.2.0
scikit-learn==1.3.0
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1